pyproject.toml
[tool.poetry]
name = "crypto_proj_v2.1"
version = "0.1.0"
description = ""
authors = ["omanduj <omanduj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
